refactor locale to localization
